backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.30.0
psycopg2-binary==2.9.9
redis==5.1.1
httpx==0.27.2
pyyaml==6.0.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
websockets==13.1
docker==7.1.0
gitpython==3.1.43
celery==5.4.0
prometheus-client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
croniter==3.0.3
